Add parse-diff util
