+++ lib/pull-request-handler.js
@@ -1,5 +1,5 @@
 const handlerSetup = require('./utils/handler-setup')
-const generateAssignedTo = require('./utils/generate-assigned-to')
+const parseDiff = require('./utils/parse-diff')
 const { truncate } = require('./utils/helpers')
 const { comment } = require('./templates')
 
@@ -19,14 +19,9 @@ module.exports = async context => {
   let match
   while ((match = regex.exec(diff)) !== null) {
     context.log('Match found')
-    const {
-      1: filename,
-      3: contents,
-      4: matchedLine,
-      5: prefix,
-      6: keyword,
-      7: title
-    } = match
+
+    const parsed = parseDiff({ match, context, config })
+    const title = truncate(parsed.title)
 
     // This PR already has a comment for this item
     if (comments.some(c => c.body.startsWith(`## ${title}`))) {
@@ -34,29 +29,14 @@ module.exports = async context => {
       continue
     }
 
-    const lineNumber = parseInt(match[2])
-
-    context.log({ filename, lineNumber, matchedLine, prefix, keyword, title })
-
-    const addedLines = contents.split(/^-.*\n?/m).join('')
-    const beforeMatch = addedLines.substring(0, addedLines.indexOf('\n' + matchedLine))
-
-    // Where the diff starts, plus the number of lines until the matched line
-    // minus one because `contents` starts with a \n
-    const start = lineNumber + beforeMatch.split(/\r\n|\r|\n/).length - 1
-
-    const { sha, user } = context.payload.pull_request.head
-    const { number } = context.payload.pull_request
-
-    const assignedToString = generateAssignedTo(config.autoAssign, user.login, number)
     const body = comment(context.repo({
-      title: truncate(title),
-      sha,
-      assignedToString,
-      number,
-      start,
-      filename,
-      keyword
+      title,
+      sha: parsed.sha,
+      assignedToString: parsed.assignedToString,
+      number: parsed.number,
+      start: parsed.start,
+      filename: parsed.filename,
+      keyword: parsed.keyword
     }))
 
     await context.github.issues.createComment(context.issue({ body }))
